README.md
# Simple Todo App

**Description:**
A minimalist to-do list application built with Flutter, designed to help users stay organized and productive. This app allows users to create, edit, and delete tasks, as well as mark tasks as completed.

**Features:**

* Add, edit, and delete tasks
* Mark tasks as completed
* Simple and intuitive user interface
* Customizable theme

**Technologies Used:**

* Flutter
* Dart
* Provider (for state management)

**Getting Started:**

1. Clone the repository: `git clone https://github.com/your-username/todo-app.git`
2. Install dependencies: `flutter pub get`
3. Run the app: `flutter run`

**Contributing:**
Feel free to contribute to this project by:
* Reporting bugs
* Suggesting new features
* Submitting pull requests

**License:**
This project is licensed under the MIT License - see the LICENSE file for details.